Extract RestrictionRelations feature from GraphProjector
