+++ Neos.ContentGraph.DoctrineDbalAdapter/Classes/Domain/Projection/GraphProjector.php
@@ -16,6 +16,7 @@ namespace Neos\ContentGraph\DoctrineDbalAdapter\Domain\Projection;
 use Doctrine\DBAL\Connection;
 use Neos\Cache\Frontend\VariableFrontend;
 use Neos\ContentGraph\DoctrineDbalAdapter\Domain\Projection\Feature\NodeRemoval;
+use Neos\ContentGraph\DoctrineDbalAdapter\Domain\Projection\Feature\RestrictionRelations;
 use Neos\ContentGraph\DoctrineDbalAdapter\Domain\Repository\ProjectionContentGraph;
 use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\CopyableAcrossContentStreamsInterface;
 use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Event;
@@ -51,6 +52,7 @@ use Neos\Flow\Annotations as Flow;
  */
 class GraphProjector implements ProjectorInterface, AfterInvokeInterface
 {
+    use RestrictionRelations;
     use NodeRemoval;
 
     const RELATION_DEFAULT_OFFSET = 128;
@@ -723,133 +725,6 @@ insert ignore into neos_contentgraph_restrictionrelation
         });
     }
 
-    /**
-     * @param ContentStreamIdentifier $contentStreamIdentifier
-     * @param NodeAggregateIdentifier $originNodeAggregateIdentifier
-     * @param DimensionSpacePointSet $affectedDimensionSpacePoints
-     * @throws \Doctrine\DBAL\DBALException
-     */
-    private function removeOutgoingRestrictionRelationsOfNodeAggregateInDimensionSpacePoints(
-        ContentStreamIdentifier $contentStreamIdentifier,
-        NodeAggregateIdentifier $originNodeAggregateIdentifier,
-        DimensionSpacePointSet $affectedDimensionSpacePoints
-    ): void {
-        $this->getDatabaseConnection()->executeUpdate('
-                -- GraphProjector::removeOutgoingRestrictionRelationsOfNodeAggregateInDimensionSpacePoints
- 
-                DELETE r.*
-                    FROM neos_contentgraph_restrictionrelation r
-                    WHERE r.contentstreamidentifier = :contentStreamIdentifier
-                    AND r.originnodeaggregateidentifier = :originNodeAggregateIdentifier
-                    AND r.dimensionspacepointhash in (:dimensionSpacePointHashes)',
-            [
-                'contentStreamIdentifier' => (string)$contentStreamIdentifier,
-                'originNodeAggregateIdentifier' => (string)$originNodeAggregateIdentifier,
-                'dimensionSpacePointHashes' => $affectedDimensionSpacePoints->getPointHashes()
-            ],
-            [
-                'dimensionSpacePointHashes' => Connection::PARAM_STR_ARRAY
-            ]
-        );
-    }
-
-    /**
-     * @param ContentStreamIdentifier $contentStreamIdentifier
-     * @param NodeAggregateIdentifier $nodeAggregateIdentifier
-     * @throws \Doctrine\DBAL\DBALException
-     */
-    private function removeAllRestrictionRelationsUnderneathNodeAggregate(ContentStreamIdentifier $contentStreamIdentifier, NodeAggregateIdentifier $nodeAggregateIdentifier)
-    {
-        $this->getDatabaseConnection()->executeUpdate('
-                -- GraphProjector::removeAllRestrictionRelationsUnderneathNodeAggregate
- 
-                delete r.* from
-                    neos_contentgraph_restrictionrelation r
-                    join 
-                     (
-                        -- we build a recursive tree
-                        with recursive tree as (
-                             -- --------------------------------
-                             -- INITIAL query: select the root nodes of the tree
-                             -- --------------------------------
-                             select
-                                n.relationanchorpoint,
-                                n.nodeaggregateidentifier,
-                                h.dimensionspacepointhash
-                             from
-                                neos_contentgraph_node n
-                             -- we need to join with the hierarchy relation, because we need the dimensionspacepointhash.
-                             inner join neos_contentgraph_hierarchyrelation h
-                                on h.childnodeanchor = n.relationanchorpoint
-                             where
-                                n.nodeaggregateidentifier = :entryNodeAggregateIdentifier
-                                and h.contentstreamidentifier = :contentStreamIdentifier
-                        union
-                             -- --------------------------------
-                             -- RECURSIVE query: do one "child" query step
-                             -- --------------------------------
-                             select
-                                c.relationanchorpoint,
-                                c.nodeaggregateidentifier,
-                                h.dimensionspacepointhash
-                             from
-                                tree p
-                             inner join neos_contentgraph_hierarchyrelation h
-                                on h.parentnodeanchor = p.relationanchorpoint
-                             inner join neos_contentgraph_node c
-                                on h.childnodeanchor = c.relationanchorpoint
-                             where
-                                h.contentstreamidentifier = :contentStreamIdentifier
-                        )
-                        select * from tree
-                     ) as tree
-
-                -- the "tree" CTE now contains a list of tuples (nodeAggregateIdentifier,dimensionSpacePointHash)
-                -- which are *descendants* of the starting NodeAggregateIdentifier in ALL DimensionSpacePointHashes
-                where
-                    r.contentstreamidentifier = :contentStreamIdentifier
-                    and r.dimensionspacepointhash = tree.dimensionspacepointhash
-                    and r.affectednodeaggregateidentifier = tree.nodeaggregateidentifier
-            ',
-            [
-                'entryNodeAggregateIdentifier' => (string)$nodeAggregateIdentifier,
-                'contentStreamIdentifier' => (string)$contentStreamIdentifier,
-            ]);
-    }
-
-    /**
-     * @param ContentStreamIdentifier $contentStreamIdentifier
-     * @param NodeAggregateIdentifier $entryNodeAggregateIdentifier
-     * @param DimensionSpacePointSet $affectedDimensionSpacePoints
-     * @throws \Doctrine\DBAL\DBALException
-     */
-    private function removeAllRestrictionRelationsInSubtreeImposedByAncestors(
-        ContentStreamIdentifier $contentStreamIdentifier,
-        NodeAggregateIdentifier $entryNodeAggregateIdentifier,
-        DimensionSpacePointSet $affectedDimensionSpacePoints
-    ) {
-        $descendantNodeAggregateIdentifiers = $this->projectionContentGraph->findDescendantNodeAggregateIdentifiers($contentStreamIdentifier, $entryNodeAggregateIdentifier, $affectedDimensionSpacePoints);
-
-        $this->getDatabaseConnection()->executeUpdate('
-                -- GraphProjector::removeAllRestrictionRelationsInSubtreeImposedByAncestors
- 
-                DELETE r.*
-                    FROM neos_contentgraph_restrictionrelation r
-                    WHERE r.contentstreamidentifier = :contentStreamIdentifier
-                    AND r.originnodeaggregateidentifier NOT IN (:descendantNodeAggregateIdentifiers)
-                    AND r.affectednodeaggregateidentifier IN (:descendantNodeAggregateIdentifiers)
-                    AND r.dimensionspacepointhash IN (:affectedDimensionSpacePointHashes)',
-            [
-                'contentStreamIdentifier' => (string)$contentStreamIdentifier,
-                'descendantNodeAggregateIdentifiers' => array_keys($descendantNodeAggregateIdentifiers),
-                'affectedDimensionSpacePointHashes' => $affectedDimensionSpacePoints->getPointHashes()
-            ],
-            [
-                'descendantNodeAggregateIdentifiers' => Connection::PARAM_STR_ARRAY,
-                'affectedDimensionSpacePointHashes' => Connection::PARAM_STR_ARRAY
-            ]
-        );
-    }
 
     /**
      * @param NodeSpecializationVariantWasCreated $event
